requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
supabase>=2.0.0
python-dotenv>=1.0.0  # 环境变量管理